requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI
openai>=1.0.0
pydub
ffmpg

